Restructure the academic program page components
