+++ src/components/AcademicPreparationPage/ExamCategories/ExamCategories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { Card, CardHeader, CardTitle, CardContent, CardFooter } from "../ui/card";
-import { Button } from "../ui/button";
-import { Badge } from "../ui/badge";
+import { Card, CardHeader, CardTitle, CardContent, CardFooter } from "../../ui/card";
+import { Button } from "../../ui/button";
+import { Badge } from "../../ui/badge";
 
 const ExamCategories: React.FC = () => {
   const categories = [
