+++ src/components/AcademicPreparationPage/FAQ/FAQ.tsx
@@ -4,7 +4,7 @@ import {
   AccordionContent,
   AccordionItem,
   AccordionTrigger,
-} from "../ui/accordion";
+} from "../../ui/accordion"
 
 interface FAQItem {
   question: string;
