+++ src/components/AcademicPreparationPage/HeroSection/HeroSection.tsx
