+++ src/pages/sections/AcademicPreparation/AcademicPreparation.tsx
@@ -1,11 +1,11 @@
 
-import Slider from "../../../components/Corousel/Slider"
-import CourseCarousel from "../../../components/CourseCarousel/CourseCarousel"
-import ExamCategories from "../../..//components/ExamCategories/ExamCategories"
-import FAQ from "../../../components/FAQ/FAQ"
-import HeroSection from "../../../components/HeroSection/HeroSection"
-import OnlineClassesHero from "../../../components/OnlineClassesHero/OnlineClassesHero"
-import TrustedPlatform from "../../../components/TrustedPlatform/TrustedPlatform"
+import Slider from "../../../components/AcademicPreparationPage/Corousel/Slider"
+import CourseCarousel from "../../../components/AcademicPreparationPage/CourseCarousel/CourseCarousel"
+import ExamCategories from "../../../components/AcademicPreparationPage/ExamCategories/ExamCategories"
+import FAQ from "../../../components/AcademicPreparationPage/FAQ/FAQ"
+import HeroSection from "../../../components/AcademicPreparationPage/HeroSection/HeroSection"
+import OnlineClassesHero from "../../../components/AcademicPreparationPage/OnlineClassesHero/OnlineClassesHero"
+import TrustedPlatform from "../../../components/AcademicPreparationPage/TrustedPlatform/TrustedPlatform"
 import Header from "../../../components/Header"
 import Footer from "../../../components/Footer"
